Fix subscription update adds (#200)
* fix horse detail page map location

* Added keywords as None

* Fix distinct query from search horse

* Update font sizing on landing page

* Fix onesignal notification creation

* search-by-name update to look for keywords

* update the search criteria for home search icon

* update search horse criteria and fav with like

* Update maps

* update url

* added unread-count-notification

* Fix issues

* fix location points

* fix lat lng

* update lat lng on backend

* update lat lng on frontend

* fix horse location issue

* Fix search horse random response and search horse name

* Added google geocoding api

* Fix subscription update adds

+++ backend/users/models.py
@@ -58,7 +58,7 @@ class UserProfile(models.Model):
     subscription_start_date = models.DateTimeField(null=True, blank=True)
     subscription_renew_date = models.DateTimeField(null=True, blank=True)
     promotion_adds = models.IntegerField(default=5)
-    subscription_adds = models.IntegerField(default=0)
+    subscription_adds = models.BigIntegerField(default=0)
 
     ban_user_from_posting = models.BooleanField(default=False)
     ban_user_from_posting_date = models.DateField(null=True, blank=True)
